+++ examples/simd_stat/stat_float_variance.cxx
@@ -12,7 +12,7 @@ int main()
   std::cout << "\nVariance (Float) Example." << std::endl;
 
   // Sequence of values
-  std::vector<double> x{ 0.2, 0.4, 0.6, 0.8, 1.0 };
+  std::vector<float> x{ 0.2, 0.4, 0.6, 0.8, 1.0 };
 
   // Length of sequence
   std::int32_t n = x.size();
@@ -23,8 +23,8 @@ int main()
   //double m3 = nola::variance_avx512( n, x.data() );
 
   // Display result
-  std::cout << "\nvariance (serial) = " << m1 // m1 = 0.6000
-            //<< "\nvariance (avx2)   = " << m2 // m2 = 0.6000
-            //<< "\nvariance (avx512) = " << m3 // m3 = 0.6000
+  std::cout << "\nvariance (serial) = " << m1 // m1 = 0.1000
+            //<< "\nvariance (avx2)   = " << m2 // m2 = 0.1000
+            //<< "\nvariance (avx512) = " << m3 // m3 = 0.1000
             << std::endl;
 }
